+++ src/components/chart/o-chart-data-adapter.factory.ts
@@ -1,19 +1,11 @@
 import { ChartDataAdapterFactory, ChartDataAdapter } from '../../interfaces';
 
 import {
-  ChartConfiguration,
-  GenericDataAdapter,
-  PieDataAdapter
+  ChartConfiguration, GenericDataAdapter, PieDataAdapter, GaugeDataAdapter, LineDataAdapter,
+  ScatterDataAdapter, DiscreteBarDataAdapter, MultiBarDataAdapter, MultiBarHorizontalDataAdapter,
+  LinePlusBarDataAdapter, ForceDirectedGraphDataAdapter, CandlestickDataAdapter, OhlcDataAdapter,
+  BulletDataAdapter, GaugeSimpleDataAdapter, StackedAreaDataAdapter, RadarDataAdapter, ParallelCoordinatesDataAdapter
 } from '../../core';
-import { LineDataAdapter } from '../../core/data-adapters/line-data-adapter';
-import { ScatterDataAdapter } from '../../core/data-adapters/scatter-data-adapter';
-import { DiscreteBarDataAdapter } from '../../core/data-adapters/discreteBar-data-adapter';
-import { MultiBarDataAdapter } from '../../core/data-adapters/multibar-data-adapter';
-import { MultiBarHorizontalDataAdapter } from '../../core/data-adapters/multibar-horizontal-data-adapter';
-import { LinePlusBarDataAdapter } from '../../core/data-adapters/lineplusbar-data-adapter';
-import { ForceDirectedGraphDataAdapter } from '../../core/data-adapters/forceDirectedGraph-data-adapter';
-import { CandlestickDataAdapter } from '../../core/data-adapters/candlestick-data-adapter';
-import { OhlcDataAdapter } from '../../core/data-adapters/ohlc-data-adapter';
 
 export class OChartDataAdapterFactory implements ChartDataAdapterFactory {
 
@@ -58,6 +50,27 @@ export class OChartDataAdapterFactory implements ChartDataAdapterFactory {
         break;
       case 'ohlcBarChart':
         adapter = this.createOhlcDataAdapter();
+        break;
+      case 'bulletChart':
+        adapter = this.createBulletDataAdapter();
+        break;
+      case 'gaugeDashboardChart':
+      case 'gaugeSlimChart':
+      case 'gaugeSpaceChart':
+      case 'radialPercentChart':
+        adapter = this.createGaugeDataAdapter();
+        break;
+      case 'gaugeSimpleChart':
+        adapter = this.createGaugeSimpleDataAdapter();
+        break;
+      case 'stackedAreaChart':
+        adapter = this.createStackedAreaDataAdapter();
+        break;
+      case 'radarChart':
+        adapter = this.createRadarDataAdapter();
+        break;
+      case 'parallelCoordinatesChart':
+        adapter = this.createParallelCoordinatesDataAdapter();
         break;
     }
     return adapter;
@@ -105,5 +118,29 @@ export class OChartDataAdapterFactory implements ChartDataAdapterFactory {
 
   public createOhlcDataAdapter(): ChartDataAdapter {
     return new OhlcDataAdapter(this.chartConf);
+  }
+
+  public createBulletDataAdapter(): ChartDataAdapter {
+    return new BulletDataAdapter(this.chartConf);
+  }
+
+  public createGaugeDataAdapter(): ChartDataAdapter {
+    return new GaugeDataAdapter(this.chartConf);
+  }
+
+  public createGaugeSimpleDataAdapter(): ChartDataAdapter {
+    return new GaugeSimpleDataAdapter(this.chartConf);
+  }
+
+  public createStackedAreaDataAdapter(): ChartDataAdapter {
+    return new StackedAreaDataAdapter(this.chartConf);
+  }
+
+  public createRadarDataAdapter(): ChartDataAdapter {
+    return new RadarDataAdapter(this.chartConf);
+  }
+
+  public createParallelCoordinatesDataAdapter(): ChartDataAdapter {
+    return new ParallelCoordinatesDataAdapter(this.chartConf);
   }
 }