+++ src/core/chart-options/DiscreteBarChartConfiguration.class.ts
@@ -4,6 +4,8 @@ import { ChartAxisConfiguration } from './ChartAxisConfiguration.class';
 import { ChartMarginConfiguration } from './ChartMarginConfiguration.class';
 
 export class DiscreteBarChartConfiguration extends ChartConfiguration {
+  readonly type: string = 'discreteBar';
+
   showValues: boolean = true;
   duration: number = 500;
 
@@ -17,8 +19,8 @@ export class DiscreteBarChartConfiguration extends ChartConfiguration {
   showXAxis: boolean = true;
   showYAxis: boolean = true;
   rightAlingYAxis: boolean = false;
-  xAxis1: ChartAxisConfiguration;
-  yAxis1: ChartAxisConfiguration;
+  x1Axis: ChartAxisConfiguration;
+  y1Axis: ChartAxisConfiguration;
 
   margin: ChartMarginConfiguration;
 
@@ -28,8 +30,8 @@ export class DiscreteBarChartConfiguration extends ChartConfiguration {
   constructor() {
     super();
     this.legend = new ChartLegendConfiguration();
-    this.xAxis1 = new ChartAxisConfiguration();
-    this.yAxis1 = new ChartAxisConfiguration();
+    this.x1Axis = new ChartAxisConfiguration();
+    this.y1Axis = new ChartAxisConfiguration();
     this.margin = new ChartMarginConfiguration();
   }
 }