+++ src/core/chart-options/ForceDirectedGraphOptions.class.ts
@@ -10,22 +10,22 @@ export class ForceDirectedGraphOptions extends AbstractChartOptions {
   }
 
   public getChartOptions(): Object {
-    let params = this.chartConf as ForceDirectedGraphConfiguration;
+    let conf = this.chartConf instanceof ForceDirectedGraphConfiguration ? this.chartConf : new ForceDirectedGraphConfiguration();
 
     let chart = {
       type: this.getChartType(),
-      height: params.height,
-      width: params.width,
+      height: conf.height,
+      width: conf.width,
       color: this.getColor(),
-      margin: params.margin.getMarginOptions(),
-      linkStrenght: params.linkStrength,
-      friction: params.friction,
-      linkDist: params.linkDist,
-      charge: params.charge,
-      gravity: params.gravity,
-      theta: params.theta,
-      alpha: params.alpha,
-      radius: params.radius
+      margin: conf.margin.getMarginOptions(),
+      linkStrenght: conf.linkStrength,
+      friction: conf.friction,
+      linkDist: conf.linkDist,
+      charge: conf.charge,
+      gravity: conf.gravity,
+      theta: conf.theta,
+      alpha: conf.alpha,
+      radius: conf.radius
     };
     return chart;
   }